Rename NorentConfirmationModal to YesNoConfirmationModal.
